Move SimpleIndexingAnnotations to .internal package
Also, replace system property checks with static enable/disable flag.
