+++ document/abi-spec.json
@@ -1357,27 +1357,6 @@
     ],
     "fields" : [ ]
   },
-  "com.yahoo.document.annotation.SimpleIndexingAnnotations" : {
-    "superClass" : "java.lang.Object",
-    "interfaces" : [ ],
-    "attributes" : [
-      "public",
-      "final"
-    ],
-    "methods" : [
-      "public void <init>()",
-      "public void add(int, int, java.lang.String)",
-      "public boolean equals(java.lang.Object)",
-      "public int getCount()",
-      "public int getFrom(int)",
-      "public int getLength(int)",
-      "public java.lang.String getTerm(int)",
-      "public int hashCode()",
-      "public com.yahoo.document.annotation.SpanTree toSpanTree(java.lang.String)",
-      "public java.lang.String toString()"
-    ],
-    "fields" : [ ]
-  },
   "com.yahoo.document.annotation.Span" : {
     "superClass" : "com.yahoo.document.annotation.SpanNode",
     "interfaces" : [ ],
@@ -2196,9 +2175,9 @@
       "public com.yahoo.document.annotation.SpanTree getSpanTree(java.lang.String)",
       "public com.yahoo.document.annotation.SpanTree setSpanTree(com.yahoo.document.annotation.SpanTree)",
       "public com.yahoo.document.annotation.SpanTree removeSpanTree(java.lang.String)",
-      "public com.yahoo.document.annotation.SimpleIndexingAnnotations createSimpleAnnotations()",
-      "public com.yahoo.document.annotation.SimpleIndexingAnnotations getSimpleAnnotations()",
-      "public void setSimpleAnnotations(com.yahoo.document.annotation.SimpleIndexingAnnotations)",
+      "public com.yahoo.document.annotation.internal.SimpleIndexingAnnotations createSimpleAnnotations()",
+      "public com.yahoo.document.annotation.internal.SimpleIndexingAnnotations getSimpleAnnotations()",
+      "public void setSimpleAnnotations(com.yahoo.document.annotation.internal.SimpleIndexingAnnotations)",
       "public java.lang.String getString()",
       "public java.lang.Object getWrappedValue()",
       "public void printXml(com.yahoo.document.serialization.XmlStream)",
