+++ document/src/main/java/com/yahoo/document/annotation/internal/package-info.java
@@ -1,0 +1,9 @@
+// Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+/**
+ * Internal implementation details for annotation optimization.
+ * NOT part of public API - for internal use only.
+ */
+@ExportPackage
+package com.yahoo.document.annotation.internal;
+
+import com.yahoo.osgi.annotation.ExportPackage;
