+++ indexinglanguage/src/test/java/com/yahoo/vespa/indexinglanguage/linguistics/SimpleAnnotationsComparisonTestCase.java
@@ -3,7 +3,7 @@ package com.yahoo.vespa.indexinglanguage.linguistics;
 
 import com.yahoo.document.annotation.Annotation;
 import com.yahoo.document.annotation.AnnotationTypes;
-import com.yahoo.document.annotation.SimpleIndexingAnnotations;
+import com.yahoo.document.annotation.internal.SimpleIndexingAnnotations;
 import com.yahoo.document.annotation.Span;
 import com.yahoo.document.annotation.SpanTree;
 import com.yahoo.document.annotation.SpanTrees;
